Moving class to extension package.
